+++ src/packet/smi.rs
@@ -4,7 +4,8 @@
 
 use core::fmt;
 
-use super::{payload, unit, Decode, Decoder, Error};
+use super::decoder::{Decode, Decoder};
+use super::{payload, unit, Error};
 
 /// A Siemens Messaging Infrastructure (SMI) Packet
 ///